requirements.txt
dankware
mcstatus
packaging
psutil
pyminizip
pynput
pypresence
python-dateutil
requests
rich
unitypackff
win11toast
translatepy
playsound==1.2.2